requirements.txt
scipy==1.6.3
lxml==4.6.3
contextlib2==0.6.0.post1
numpy==1.19.5
absl_py==0.12.0
tensorflow==1.15.2
six==1.15.0
streamlit==0.81.1
matplotlib==3.4.2
pandas==1.2.4
opencv_python==4.5.2.52

apache_beam==2.29.0
ipython==7.24.0
lvis==0.5.3
mock==4.0.3
nets==0.0.3.1
Pillow==8.2.0
protobuf==3.17.1
pycocotools==2.0.2
tf_slim==1.1.0
tensorflow-object-detection-api==0.1.1
